requirements.txt
# Nutrition Agent System - Dependencies

# Core ADK Framework
google-adk>=0.1.0

# Google Generative AI
google-generativeai>=0.3.0

# Environment Management
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0

# Database
# sqlite3 is built-in with Python, no installation needed

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP & API (for future integrations)
requests>=2.31.0
fastapi>=0.104.0  # If building REST API
uvicorn>=0.24.0   # ASGI server

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development Tools
black>=23.0.0      # Code formatting
flake8>=6.1.0      # Linting
mypy>=1.6.0        # Type checking

# Optional: Enhanced Features
# jupyter>=1.0.0   # For notebook integration
# streamlit>=1.28.0  # For web UI
# plotly>=5.17.0   # For visualization
